+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.0.13...HEAD) - TBD
 
+### :bug: Fixed
+
+- Negative test cases with invalid parameters mistakenly becoming positive test cases. [#2900](https://github.com/schemathesis/schemathesis/issues/2900), [#2913](https://github.com/schemathesis/schemathesis/issues/2913)
+
 ## [4.0.13](https://github.com/schemathesis/schemathesis/compare/v4.0.12...v4.0.13) - 2025-07-24
 
 ### :bug: Fixed